pyproject.toml
[project]
name = "meta-sam3-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "geopy>=2.4.1",
    "huggingface-hub>=1.8.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "requests>=2.33.0",
    "shapely>=2.1.2",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "transformers>=5.4.0",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu128"

[tool.uv]
index-strategy = "unsafe-best-match"
